backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
psycopg2-binary
httpx
pinecone
python-dotenv
pillow
numpy
python-multipart
